requirements.txt
requests
Pillow
beautifulsoup4
matplotlib
numpy
pandas
netCDF4
opencv-python-headless
pytesseract
urllib3
